Refined focuser/rotator behaviour

+++ src/lib/FocuserDC.h
@@ -114,7 +114,7 @@ class focuserDC {
 
     // get position
     double getPosition() {
-      if ((target.part.m == spos) && (abs(spos - round(requestedPosition/spm)) <= spm)) return requestedPosition; else return ((double)spos)/spm;
+      if (((long)target.part.m == spos) && (abs(spos - round(requestedPosition/spm)) <= spm)) return requestedPosition; else return ((double)spos)/spm;
     }
 
     // sets current position in microns
@@ -122,13 +122,13 @@ class focuserDC {
       spos=round(pos*spm);
       if (spos < smin) spos=smin; if (spos > smax) spos=smax;
       target.part.m=spos; target.part.f=0;
-      requestedPosition=((double)target.part.m)/spm;
+      requestedPosition=((long)target.part.m)/spm;
     }
 
     // sets target position in microns
     void setTarget(double pos) {
       dcMotor.setPower((moveRate/1000.0)*powerFor1mmSec);
-      target.part.m=round(pos*spm); target.part.f=0;
+      target.part.m=(long)round(pos*spm); target.part.f=0;
       if ((long)target.part.m < smin) target.part.m=smin; if ((long)target.part.m > smax) target.part.m=smax;
       requestedPosition=pos;
     }
@@ -136,16 +136,16 @@ class focuserDC {
     // sets target relative position in microns
     void relativeTarget(double pos) {
       dcMotor.setPower((moveRate/1000.0)*powerFor1mmSec);
-      target.part.m+=round(pos*spm); target.part.f=0;
+      target.part.m+=(long)round(pos*spm); target.part.f=0;
       if ((long)target.part.m < smin) target.part.m=smin; if ((long)target.part.m > smax) target.part.m=smax;
-      requestedPosition=((double)target.part.m)/spm;
+      requestedPosition=((long)target.part.m)/spm;
     }
 
     // do automatic movement
     void move() {
       target.fixed+=delta.fixed;
       // update requested position
-      if (delta.fixed != 0) requestedPosition=((double)target.part.m)/spm;
+      if (delta.fixed != 0) requestedPosition=((long)target.part.m)/spm;
       // stop at limits
       if (((long)target.part.m < smin) || ((long)target.part.m > smax)) delta.fixed=0;
     }
